+++ packages/blocks-core/CHANGELOG.md
@@ -37,6 +37,7 @@
 - `Navigation Link` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Navigation Submenu` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Post Content` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Post Navigation Link` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests
@@ -75,6 +76,7 @@
 - Added E2E tests to check functionality `Navigation Link` new inner block.
 - Added E2E tests to check functionality `Navigation Submenu` new inner block.
 - Added E2E tests to check functionality `Post Content` new inner block.
+- Added E2E tests to check functionality `Post Navigation Link` new inner block.
 
 ## 1.4.0 (2025-05-19)